Fix bug of players with null total games

+++ src/main/java/com/mafia/statistics/MafiaStatisticsAPI/dto/statistics/additional/TopGamesTable.java
@@ -17,5 +17,5 @@ public class TopGamesTable {
 
     private String playerNickname;
 
-    private Integer gamesTotal;
+    private Long gamesTotal;
 }
